PIM-5384: move export method in proper class
