+++ features/Pim/Behat/Context/JobContext.php
@@ -1,0 +1,118 @@
+<?php
+
+namespace Pim\Behat\Context;
+
+use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
+use Behat\Behat\Context\Step;
+use Behat\Gherkin\Node\TableNode;
+
+class JobContext extends PimContext
+{
+    /**
+     * @param string    $code
+     * @param TableNode $table
+     *
+     * @Given /^the following job "([^"]*)" configuration:$/
+     */
+    public function theFollowingJobConfiguration($code, TableNode $table)
+    {
+        $jobInstance   = $this->getFixturesContext()->getJobInstance($code);
+        $configuration = $jobInstance->getRawConfiguration();
+
+        foreach ($table->getRowsHash() as $property => $value) {
+            $value = $this->replacePlaceholders($value);
+            if (in_array($value, ['yes', 'no'])) {
+                $value = 'yes' === $value;
+            }
+
+            $configuration[$property] = $value;
+        }
+
+        $jobInstance->setRawConfiguration($configuration);
+        // TODO use a Saver
+        $this->getFixturesContext()->flush();
+    }
+
+    /**
+     * @param string $type
+     *
+     * @When /^I launch the (import|export) job$/
+     */
+    public function iExecuteTheJob($type)
+    {
+        $this->getPage(sprintf('%s show', ucfirst($type)))->execute();
+    }
+
+    /**
+     * @param string $file
+     * @param string $field
+     *
+     * @Given /^I attach file "([^"]*)" to "([^"]*)"$/
+     */
+    public function attachFileToField($file, $field)
+    {
+        if ($this->getMinkParameter('files_path')) {
+            $fullPath = rtrim(realpath($this->getMinkParameter('files_path')), DIRECTORY_SEPARATOR)
+                .DIRECTORY_SEPARATOR.$file;
+            if (is_file($fullPath)) {
+                $file = $fullPath;
+            }
+        }
+
+        $this->getCurrentPage()->attachFileToField($field, $file);
+        $this->getMainContext()->wait();
+    }
+
+    /**
+     * @param string      $action
+     * @param JobInstance $job
+     *
+     * @return \Behat\Behat\Context\Step\Then
+     *
+     * @When /^I should not be able to (launch|edit) the ("([^"]*)" (export|import) job)$/
+     */
+    public function iShouldNotBeAbleToAccessTheJob($action, JobInstance $job)
+    {
+        $this->currentPage = sprintf("%s %s", ucfirst($job->getType()), $action);
+        $page              = $this->getCurrentPage()->open(['id' => $job->getId()]);
+
+        return new Step\Then('I should see "403 Forbidden"');
+    }
+
+    /**
+     * @param string $jobType
+     *
+     * @Given /^I try to create an unknown (import|export)$/
+     */
+    public function iTryToCreateAnUnknownJob($jobType)
+    {
+        $this->getNavigationContext()->openPage(sprintf('%s creation', ucfirst($jobType)));
+    }
+
+    /**
+     * @param string $code
+     *
+     * @return string
+     */
+    public function getJobInstancePath($code)
+    {
+        $jobInstance   = $this->getFixturesContext()->getJobInstance($code);
+        $configuration = $this->getFixturesContext()->findEntity('JobConfiguration', [
+            'jobExecution' => $jobInstance->getJobExecutions()->first()
+        ]);
+
+        $step = $this->getMainContext()
+            ->getContainer()
+            ->get('akeneo_batch.connectors')
+            ->getJob($jobInstance)
+            ->getSteps()[0];
+
+        $context = json_decode(stripcslashes($configuration->getConfiguration()), true);
+
+        if (null !== $context) {
+            $step->setConfiguration($context);
+        }
+
+        return $step->getWriter()->getPath();
+    }
+}
